restored white background

+++ oculus_ros2/src/sonar_viewer.cpp
@@ -115,7 +115,7 @@ void SonarViewer::publishFan(const int& width,
 
   cv::Mat out = cv::Mat::ones(cv::Size(image_width, height), CV_MAKETYPE(mat_encoding, 1)) * std::numeric_limits<uint8_t>::max();
   cv::remap(sonar_mat_data.t(), out, source_map_1, source_map_2, cv::INTER_CUBIC, cv::BORDER_CONSTANT,
-      cv::Scalar(std::numeric_limits<uint8_t>::max(), std::numeric_limits<uint8_t>::max(), std::numeric_limits<uint8_t>::max()));
+      cv::Scalar(std::numeric_limits<uint16_t>::max(), std::numeric_limits<uint16_t>::max(), std::numeric_limits<uint16_t>::max()));
 
   // Publish sonar conic image
   sensor_msgs::msg::Image msg;
